+++ src/main/java/com/bookspot/users/presentation/UserAuthController.java
@@ -1,7 +1,7 @@
 package com.bookspot.users.presentation;
 
 import com.bookspot.users.infra.GoogleTokenVerifier;
-import com.bookspot.users.infra.JwtProvider;
+import com.bookspot.global.auth.JwtProvider;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
